sam+RL/requirements.txt
# SAM2 相关
torch>=2.5.1
torchvision>=0.20.1
numpy<2.0
opencv-python
Pillow

# RL 相关
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# 血管分析相关
scikit-image
scipy
scikit-learn<1.4.0

# 可视化与日志
matplotlib
tensorboard
tqdm
pyyaml

# 其他
einops

